+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '@/components/App/App';
+import App from '@/containers/App';
 import * as serviceWorker from './utils/serviceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
