--- src/routes/Styleguide/index.tsx
@@ -1,3 +1,0 @@
-import StyleGuide from './Styleguide';
-
-export default StyleGuide;
