+++ src/routes/index.tsx
@@ -4,5 +4,4 @@ export { default as NotFound } from './NotFound';
 export { default as ResetPassword } from './ResetPassword';
 export { default as SignIn } from './SignIn';
 export { default as SignUp } from './SignUp';
-export { default as Styleguide } from './Styleguide';
 export { default as TierLists } from './TierLists';
