+++ src/theme/index.tsx
@@ -1,0 +1,3 @@
+export { lightTheme } from './lightTheme';
+export { darkTheme } from './darkTheme';
+export { getTextColorByBackground } from './utils';
